+++ src/console/controllers/CacheController.php
@@ -335,6 +335,24 @@ class CacheController extends Controller
     }
 
     /**
+     * Refreshes cached pages that track expired elements.
+     *
+     * @since 4.15.0
+     */
+    public function actionRefreshExpiredElements(): int
+    {
+        Blitz::$plugin->refreshCache->refreshExpiredElements();
+
+        if (!$this->queue) {
+            Craft::$app->runAction('queue/run');
+        }
+
+        $this->output('Expired elements successfully refreshed.');
+
+        return ExitCode::OK;
+    }
+
+    /**
      * Refreshes all pages in the provided site.
      */
     public function actionRefreshSite(int $siteId = null): int