+++ CHANGELOG.md
@@ -4,15 +4,22 @@
 
 ### Added
 
+- Added a new `blitz/cache/refresh-expired-elements` console command for refreshing cached pages that track expired elements.
 - Added the `BaseCachePurger::shouldPurgeAfterRefresh()` method.
 
 ### Changed
 
-- Drivers are now listed in alphabetical order, with “None” first.
+- Drivers are now listed in alphabetical order, with “None” always listed first.
 - Improved the progress indicator during the generate cache process.
-- The `Content-Encoding` header is no longer removed from responses.
 - Simplified the default cache control header settings.
+- Pages that return empty content are no longer ignored from being cached.
+- The `Content-Encoding` header is no longer removed from responses.
 - Cached and dynamic includes are now resolved as regular includes when previewing an element ([#646](https://github.com/putyourlightson/craft-blitz/issues/646)).
+
+### Fixed
+
+- Fixed a bug in which expiry dates were always being output using the current timestamp in the Blitz Diagnostics utility.
+- Fixed an edge-case bug in which cached includes were not being refreshed when expired.
 
 ## 5.0.0-beta.4 - 2024-03-25